Swap CC0-1.0 license for MIT-0
Unlike other CC licenses, CC0-1.0 was considered fine for software.
However, it appears many are concerned about its explicit patent
language. While it should not be a problem for the way it was used in
this package, I don't feel like dealing with this. So let's swap it with
the next best thing for our use case; going with MIT-0 doesn't make much
difference for us.

https://lwn.net/Articles/902410/